Added test.dat to gitignore and updated my_lib

+++ Source Code/my_lib.h
@@ -34,7 +34,7 @@ double psi(double x,int n) {
     } else {
         psi_n[1] = x_t*sqrt(2.0/1.0)*psi_n[0];
         for (int i = 2; i <= n; i++) {
-            psi_n[i] = x_t*sqrt(2.0/i)*psi_n[i-1] - sqrt((i-1.0)/double(i))*psi_n[i-2];
+            psi_n[i] = x_t*sqrt(2.0/i)*psi_n[i-1] - sqrt((i-1.0)/i)*psi_n[i-2];
         }
         return psi_n[n];
     }
